Multiple tokens on single file (#47)
* Physically delete file only if no other tokens refer to it
* Create DB indices

+++ src/Model/File.php
@@ -103,11 +103,15 @@ class File extends Model
             }
         });
 
-        // delete physical file from storage after we delete DB record
+        // delete physical file from storage after we delete DB record - only if no other tokens refer to same file
         $this->onHookShort(Model::HOOK_AFTER_DELETE, function () {
             $path = $this->get('location');
             if ($path && $this->flysystem->fileExists($path)) {
-                $this->flysystem->delete($path);
+                $files = (clone $this->getModel())->addCondition('id', '!=',
+                    $this->getId())->addCondition('location', $path);
+                if ($files->tryLoadAny() == null) {
+                    $this->flysystem->delete($path);
+                }
             }
         });
     }
